Bug fixes and enhancements to USB Relay and Neopixels
Now I can turn the neopixel strip on AND OFF from the web interface.
Also the web interface indicates when it is off.
USB Relays are not easily confused by sending multiple commands at once.
A single USB Relay object is used for the entire program to avoid
duplicating the relay list.