Remove XV11, Add Scanse Sweep and RTABMAP
RTABMAP tries to use 3D vision to find obstacles.
It can be a CPU hog, but I'm experimenting with it.
I stopped uisng the XV11. The Scanse Sweep is similar though.
I only use it for obstacle avoidance, it does not make maps well.
The XV11 was the same.

Hopefully the robot works fine even if you don't have the Scanse Sweep,
and don't want to use RTABMAP, but if not, let me know and we can work
to tweak the config to be more responsive to various robot
setups.