Fix and update the remote web server connection.
The code was crashing on some conditions due to some old bugs.
Improved the message handler code and added a lot of logging.
Converted to new prettier format.

I will probably remove some of the logging eventually,
but for now it is useful to see that it is working.

TODO: Add some more interactivity to the incoming SMS messages.