Remove unused Bluetooth Beacon scanner.
This was causing a lot of noise in the npm package install.
The Edystone node package was not maintained anyway, so it was bring in
some old junk that was causing noise in the install log.
There may also have been some security issues with the packages it was
pulling in.